// src/sanity-types.ts
export type Project = {
  _id: string;
  _type: "project";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  name?: string;
  description?: Array<{
    children?: Array<{
      marks?: Array<string>;
      text?: string;
      _type: "span";
      _key: string;
    }>;
    style?: "normal" | "h1" | "h2" | "h3" | "h4" | "h5" | "h6" | "blockquote";
    listItem?: "bullet" | "number";
    markDefs?: Array<{
      href?: string;
      _type: "link";
      _key: string;
    }>;
    level?: number;
    _type: "block";
    _key: string;
  }>;
  carousel?: Array<{
    _ref: string;
    _type: "reference";
    _weak?: boolean;
    _key: string;
    [internalGroqTypeReferenceTo]?: "picture";
  }>;
  github?: string;
  website?: string;
  stack?: "react" | "bubble" | "rust" | "cpp" | "python" | "javascript";
  tech?: Array<string>;
  section?: "recent" | "other" | "past";
  order?: number;
  slug?: string;
};
// ...
export declare const internalGroqTypeReferenceTo: unique symbol;